requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
transformers
torch
imaplib2
pydantic
openai
email-validator
httpx
streamlit
requests
